moved the scraper tables here, scraper is now using this module
